Switch serde support to an optional feature
