+++ .github/workflows/rust.yml
@@ -20,3 +20,5 @@ jobs:
       run: cargo build --verbose
     - name: Run tests
       run: cargo test --verbose
+    - name: Run tests (all features)
+      run: cargo test --verbose --all-features
