+++ Cargo.toml
@@ -14,7 +14,7 @@ rust-version = "1.56"
 [dependencies]
 enum_primitive = "0.1.1"
 lazy_static = "1.4.0"
-serde = { version = "1.0.106", features = ["derive"] }
+serde = { version = "1.0.106", features = ["derive"], optional = true }
 regex = "1.3.6"
 
 [dev-dependencies]
@@ -23,3 +23,6 @@ serde_json = "1.0.51"
 [badges]
 maintenance = { status = "passively-maintained" }
 travis-ci = { repository = "Majora320/steamid-ng", branch = "master" }
+
+[package.metadata.docs.rs]
+features = ["serde"]
